Upgrade to Jetty 7.0.0.RC3.
This also fixes the mismatch in the columns used by
PFC and Bouncer to compute the account key.

At some point PFC started using an opaque column on the users table
called uid, which for existing users is simply the username but for
new users is a random generated string.

This commit fixes that Bouncer was using username+password to generate
account keys, when really it needs to be uid+password. This went
undetected because all our test user records were from before the uid
change.

Patches 2 & 3: Tidy up miscellaneous upgrade cruft (thanks Coda).